Move Loki client and config to kinotic-domain; return Buffer directly
Relocate LokiProperties, LogQuery, and the Loki transport to kinotic-domain so
the Loki access lives in the lower shared module. Following domain's convention
(interfaces in api/services, impls in internal/api/services), the transport is
now a LokiClient interface plus DefaultLokiClient; LogQuery moves with it since
domain cannot depend on os-api. DefaultLogService injects the LokiClient
interface across the api boundary.

LogService now returns the Vert.x Buffer directly (Flux<Buffer> /
CompletableFuture<Buffer>) — the core binary converter extracts the bytes, so
the service no longer converts. Adds reactor-core to kinotic-domain and drops
vertx-web-client from kinotic-os-api (the WebClient moved).
