+++ kinotic-domain/build.gradle
@@ -15,6 +15,8 @@ dependencies {
     implementation 'io.vertx:vertx-web'
     implementation 'io.vertx:vertx-web-client'
 
+    implementation 'io.projectreactor:reactor-core'
+
     implementation 'co.elastic.clients:elasticsearch-java'
     implementation 'co.elastic.clients:elasticsearch-rest5-client'
