+++ kinotic-domain/src/main/java/org/kinotic/domain/api/config/LokiProperties.java
@@ -1,4 +1,4 @@
-package org.kinotic.os.api.config;
+package org.kinotic.domain.api.config;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
