+++ kinotic-domain/src/main/java/org/kinotic/domain/internal/api/services/DefaultLokiClient.java
@@ -1,4 +1,4 @@
-package org.kinotic.os.internal.api.services;
+package org.kinotic.domain.internal.api.services;
 
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
@@ -9,8 +9,9 @@ import io.vertx.ext.web.client.WebClient;
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
 import lombok.extern.slf4j.Slf4j;
-import org.kinotic.os.api.config.LokiProperties;
-import org.kinotic.os.api.model.LogQuery;
+import org.kinotic.domain.api.config.LokiProperties;
+import org.kinotic.domain.api.model.log.LogQuery;
+import org.kinotic.domain.api.services.LokiClient;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
 
@@ -19,13 +20,12 @@ import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 
 /**
- * Transport to the Loki HTTP API: historical {@code query_range} requests and live {@code tail} over a
- * WebSocket. Responses are returned as raw bytes for passthrough to the caller. Each request carries the
- * given organization id as the Loki {@code X-Scope-OrgID} tenant header.
+ * Vert.x-backed {@link LokiClient}: a {@link WebClient} for {@code query_range} and a {@link WebSocketClient}
+ * for the {@code tail} stream.
  */
 @Slf4j
 @Component
-public class LokiClient {
+public class DefaultLokiClient implements LokiClient {
 
     private static final String ORG_ID_HEADER = "X-Scope-OrgID";
     private static final String QUERY_RANGE_PATH = "/loki/api/v1/query_range";
@@ -36,7 +36,7 @@ public class LokiClient {
     private WebClient webClient;
     private WebSocketClient webSocketClient;
 
-    public LokiClient(Vertx vertx, LokiProperties lokiProperties) {
+    public DefaultLokiClient(Vertx vertx, LokiProperties lokiProperties) {
         this.vertx = vertx;
         this.lokiProperties = lokiProperties;
     }
@@ -57,13 +57,7 @@ public class LokiClient {
         }
     }
 
-    /**
-     * Runs a Loki {@code query_range} and returns the raw response body.
-     *
-     * @param orgId the Loki tenant ({@code X-Scope-OrgID})
-     * @param query the LogQL selector, time range, and limit
-     * @return a {@link Future} of the raw Loki response body
-     */
+    @Override
     public Future<Buffer> queryRange(String orgId, LogQuery query) {
         return webClient.getAbs(lokiProperties.getUrl() + QUERY_RANGE_PATH)
                         .addQueryParam("query", query.getQuery())
@@ -78,15 +72,8 @@ public class LokiClient {
                                         + (resp.bodyAsString() != null ? " — " + resp.bodyAsString() : "")));
     }
 
-    /**
-     * Opens a Loki {@code tail} WebSocket and emits each tail frame's raw bytes. The WebSocket is closed when
-     * the returned {@link Flux} is cancelled or completed.
-     *
-     * @param orgId the Loki tenant ({@code X-Scope-OrgID})
-     * @param query the LogQL query to follow
-     * @return a {@link Flux} of raw Loki tail frame bytes
-     */
-    public Flux<byte[]> tail(String orgId, String query) {
+    @Override
+    public Flux<Buffer> tail(String orgId, String query) {
         return Flux.create(sink -> webSocketClient.connect(tailOptions(orgId, query))
                 .onSuccess(ws -> {
                     // The Flux can be cancelled before the socket finishes opening; close it straight away.
@@ -94,7 +81,7 @@ public class LokiClient {
                         ws.close();
                         return;
                     }
-                    ws.textMessageHandler(message -> sink.next(message.getBytes(StandardCharsets.UTF_8)));
+                    ws.textMessageHandler(message -> sink.next(Buffer.buffer(message)));
                     ws.exceptionHandler(sink::error);
                     ws.closeHandler(unused -> sink.complete());
                     sink.onCancel(ws::close);
