+++ kinotic-os-api/build.gradle
@@ -8,7 +8,6 @@ dependencies {
     implementation project(':kinotic-idl')
 
     implementation 'io.vertx:vertx-core'
-    implementation 'io.vertx:vertx-web-client'
 
     implementation 'io.projectreactor:reactor-core'
