+++ kinotic-os-api/src/main/java/org/kinotic/os/internal/api/services/DefaultLogService.java
@@ -3,8 +3,9 @@ package org.kinotic.os.internal.api.services;
 import io.vertx.core.buffer.Buffer;
 import lombok.RequiredArgsConstructor;
 import org.kinotic.core.api.security.SecurityContext;
+import org.kinotic.domain.api.model.log.LogQuery;
 import org.kinotic.domain.api.security.OrganizationParticipant;
-import org.kinotic.os.api.model.LogQuery;
+import org.kinotic.domain.api.services.LokiClient;
 import org.kinotic.os.api.services.LogService;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
@@ -23,14 +24,13 @@ public class DefaultLogService implements LogService {
     private final SecurityContext securityContext;
 
     @Override
-    public Flux<byte[]> tail(String query) {
+    public Flux<Buffer> tail(String query) {
         return lokiClient.tail(requireOrganizationId(), query);
     }
 
     @Override
-    public CompletableFuture<byte[]> history(LogQuery query) {
+    public CompletableFuture<Buffer> history(LogQuery query) {
         return lokiClient.queryRange(requireOrganizationId(), query)
-                         .map(Buffer::getBytes)
                          .toCompletionStage()
                          .toCompletableFuture();
     }
